EditForm is able to show case data WIP, create new cover, save button
