+++ client/components/index.js
@@ -9,3 +9,4 @@ export {Login, Signup} from './auth-form'
 export {LandingPage} from './LandingPage'
 export {Edit} from './Edit'
 export {default as UserEditor} from './UserEditor'
+export {default as EditorForm} from './EditorForm'
